make 出来るようになった

+++ utils.c
@@ -6,13 +6,13 @@
 /*   By: jtanaka <jtanaka@student.42tokyo.jp>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/12/09 02:39:12 by jtanaka           #+#    #+#             */
-/*   Updated: 2020/12/09 02:39:22 by jtanaka          ###   ########.fr       */
+/*   Updated: 2020/12/10 02:48:56 by jtanaka          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
-void put_c_n_times(char c, long long n)
+void put_c_n_times(char c, size_t n)
 {
   if (n <= 0)
 	  return;
